Move render helpers into new item component
